+++ dart/lib/src/core/clients/socket_io/game_client.dart
@@ -24,25 +24,25 @@ class IOGameClient extends GameClient {
     socket.off(IOChannel.error.string);
     socket.off(IOChannel.gamestate.string);
     socket.off(IOChannel.lobby.string);
-    read.game(id).gameStatus = GameStatus.NotJoined;
+    read.gameFor(id).gameStatus = GameStatus.NotJoined;
     socket.disconnect();
   }
 
   @override
   Future<void> register() async {
-    read.game(id).gameStatus = GameStatus.NotJoined;
-    final assignedName = await socket
-        .call(IOChannel.register, {'name': read.game(id).playerName, 'id': id});
+    read.gameFor(id).gameStatus = GameStatus.NotJoined;
+    final assignedName = await socket.call(
+        IOChannel.register, {'name': read.gameFor(id).playerName, 'id': id});
 
-    read.game(id).playerName = assignedName as String;
-    read.game(id).gameStatus = GameStatus.NotStarted;
+    read.gameFor(id).playerName = assignedName as String;
+    read.gameFor(id).gameStatus = GameStatus.NotStarted;
     _watchState();
   }
 
   void _onLobby(Map<String, dynamic> lobby) {
     socket.on(IOChannel.lobby.string, (data) {
       final gameInfo = GameInfo.fromJson(data);
-      read.game(id).lobbyInfo = gameInfo;
+      read.gameFor(id).lobbyInfo = gameInfo;
     });
   }
 
@@ -51,11 +51,11 @@ class IOGameClient extends GameClient {
       final gameState = Game.fromJson(data as Map<String, dynamic>);
       socket.off(IOChannel.lobby.string);
       logger.fine(data);
-      read.game(id).gameState = gameState;
-      read.game(id).gameStatus = gameState.gameStatus;
+      read.gameFor(id).gameState = gameState;
+      read.gameFor(id).gameStatus = gameState.gameStatus;
     });
     socket.on(IOChannel.error.string, (data) {
-      read.game(id).gameError = GameError.fromJson(data);
+      read.gameFor(id).gameError = GameError.fromJson(data);
     });
     socket.on(IOChannel.lobby.string, (d) => _onLobby(d));
   }