+++ dart/lib/src/core/core/extensions.dart
@@ -1,6 +1,3 @@
-import 'package:riverpod/all.dart';
-
-import '../client.dart';
 import 'events.dart';
 import 'game.dart';
 import 'generic.dart';