+++ dart/lib/src/core/core/generic.dart
@@ -2,7 +2,6 @@ import 'dart:math';
 
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:riverpod/riverpod.dart';
 import 'events.dart';
 import 'game.dart';
 import 'player.dart';